Вынесла в константы размеры маркеров
