+++ js/constants.js
@@ -33,6 +33,13 @@ const CHECK_TIMES = [
   '14:00',
 ];
 
+const ICON_SIZES = {
+  mainPinWidth: 52,
+  mainPinHeight: 52,
+  pinWidth: 40,
+  pinHeight: 40,
+};
+
 const MAIN_COORDINATES = {
   lat: 35.68553153747553,
   lng: 139.75276363268588,
@@ -42,4 +49,4 @@ const ACCURACY = 5;
 
 const DEFAULT_AVATAR = 'img/muffin-grey.svg';
 
-export {ADVERTS_NUMBER, TYPES, ACCOMODATION_TYPE, MIN_ACCOMODATION_PRICES, MAX_ROOM_NUMBER, MIN_CAPACITY, CHECK_TIMES, MAIN_COORDINATES, ACCURACY, DEFAULT_AVATAR};
+export {ADVERTS_NUMBER, TYPES, ACCOMODATION_TYPE, MIN_ACCOMODATION_PRICES, MAX_ROOM_NUMBER, MIN_CAPACITY, CHECK_TIMES, ICON_SIZES, MAIN_COORDINATES, ACCURACY, DEFAULT_AVATAR};
